requirements.txt
requests==2.28.0
Flask==2.1.2
python-dotenv==0.20.0
google-cloud-dns==0.34.1
google-cloud-compute==1.3.2